refactor: rearrange test files
